test/refactor: move vote dto`s together to the same directory
* add additional units tests for vote service.
